+++ Jogo/Skeleton.h
@@ -5,7 +5,7 @@
 class Skeleton : public Enemy
 {
 private:
-	int damage;
+	float damage;
 
 public:
 	Skeleton(GraphicsManager* graphicsManager, float* dt, int id, float spriteScale, sf::Vector2f position, sf::Vector2f bodySize, std::string pathToTexture, std::string textureName, float speed, long int hp);
@@ -14,7 +14,7 @@ public:
 
 	void update();
 
-	int getDamage() { return this->damage; }
+	float getDamage() { return this->damage; }
 
 	void updateAnimation();
 };
